Improved workout controller error handling

+++ FitnessDiary/FitnessDiary/Controllers/WorkoutController.cs
@@ -1,4 +1,5 @@
-﻿using FitnessDiary.Core.Contracts;
+﻿using FitnessDiary.Core.Constants;
+using FitnessDiary.Core.Contracts;
 using FitnessDiary.Core.Models.Workout;
 using Microsoft.AspNetCore.Mvc;
 using System.Security.Claims;
@@ -8,10 +9,12 @@ namespace FitnessDiary.Controllers
     public class WorkoutController : Controller
     {
         private readonly IWorkoutService workoutService;
+        private readonly IAccountService accountService;
 
-        public WorkoutController(IWorkoutService _workoutService)
+        public WorkoutController(IWorkoutService _workoutService, IAccountService _accountService)
         {
             workoutService = _workoutService;
+            accountService = _accountService;
         }
 
         public  IActionResult CreateTamplate()
@@ -22,7 +25,11 @@ namespace FitnessDiary.Controllers
         [HttpPost]
         public async Task<IActionResult> CreateTamplate([FromBody]CreateWorkoutViewModel model)
         {
-            var userId = this.User.FindFirstValue(ClaimTypes.NameIdentifier);
+            var userId = accountService.GetById(this.User.FindFirstValue(ClaimTypes.NameIdentifier));
+            if (userId == null)
+            {
+                TempData[MessageConstant.ErrorMessage] = "Invalid user Id";
+            }
             if (!ModelState.IsValid)
             {
                 return View(model);
@@ -32,7 +39,12 @@ namespace FitnessDiary.Controllers
         }
         public async Task<IActionResult> MineTamlates()
         {
-            var userId = this.User.FindFirstValue(ClaimTypes.NameIdentifier);
+            var userId = accountService.GetById(this.User.FindFirstValue(ClaimTypes.NameIdentifier));
+            if (userId == null)
+            {
+                TempData[MessageConstant.ErrorMessage] = "Invalid user Id";
+                return RedirectToAction("Index", "Home");
+            }
 
             var workouts = await workoutService.GetMineTamplatesAsync(userId);
 
@@ -40,6 +52,12 @@ namespace FitnessDiary.Controllers
         }
         public async Task<IActionResult> EditTamplate(string Id)
         {
+            if ((await workoutService.TamplateExistsByIdAsync(Id)) == false)
+            {
+                TempData[MessageConstant.ErrorMessage] = "Invalid workout tamplate Id";
+                return RedirectToAction(nameof(MineTamlates));
+            }
+
             var workoutTamplate = await workoutService.GetTamplateById(Id);
 
             return View(workoutTamplate);
@@ -47,6 +65,11 @@ namespace FitnessDiary.Controllers
         [HttpPost]
         public async Task<IActionResult> EditTamplate(EditTamplateViewModel model)
         {
+            if ((await workoutService.TamplateExistsByIdAsync(model.Id)) == false)
+            {
+                TempData[MessageConstant.ErrorMessage] = "Invalid workout tamplate Id";
+                return RedirectToAction(nameof(MineTamlates));
+            }
             if (!ModelState.IsValid)
             {
                 return View(model);
@@ -59,6 +82,11 @@ namespace FitnessDiary.Controllers
         [HttpPost]
         public async Task<IActionResult> AddExerciseToTamplate([FromBody] AddExerciseModel model)
         {
+            if ((await workoutService.TamplateExistsByIdAsync(model.WorkoutId)) == false)
+            {
+                TempData[MessageConstant.ErrorMessage] = "Invalid workout tamplate Id";
+                return RedirectToAction(nameof(MineTamlates));
+            }
             if (!ModelState.IsValid)
             {
                 return RedirectToAction("MineTamlates");
@@ -71,12 +99,26 @@ namespace FitnessDiary.Controllers
         [HttpPost]
         public async Task<IActionResult> RemoveExerciseFromTamplate([FromBody] RemoveExerciseModel model)
         {
+            if ((await workoutService.TamplateExistsByIdAsync(model.TamplateId)) == false)
+            {
+                TempData[MessageConstant.ErrorMessage] = "Invalid workout tamplate Id";
+                return RedirectToAction(nameof(MineTamlates));
+            }
+            if (!ModelState.IsValid)
+            {
+                return RedirectToAction("MineTamlates");
+            }
             await workoutService.RemoveExerciseAsync(model.ExerciseId, model.TamplateId);
 
             return Json("success");
         }
         public async Task<IActionResult> AddToDiary(string Id)
         {
+            if ((await workoutService.TamplateExistsByIdAsync(Id)) == false)
+            {
+                TempData[MessageConstant.ErrorMessage] = "Invalid workout tamplate Id";
+                return RedirectToAction(nameof(MineTamlates));
+            }
             var tamplate = await workoutService.GetTamplateForDiaryByIdAsync(Id);
 
             return View(tamplate);
@@ -85,7 +127,12 @@ namespace FitnessDiary.Controllers
         [HttpPost]
         public async Task<IActionResult> AddToDiary(AddToDiaryViewModel model)
         {
-            var userId = this.User.FindFirstValue(ClaimTypes.NameIdentifier);
+            var userId = accountService.GetById(this.User.FindFirstValue(ClaimTypes.NameIdentifier));
+            if (userId == null)
+            {
+                TempData[MessageConstant.ErrorMessage] = "Invalid user Id";
+                return RedirectToAction("Index", "Home");
+            }
             if (!ModelState.IsValid)
             {
                 return RedirectToAction("AddToDiary", "Workout", new { Id = model.Id });
